Create lcc_classifier_tests.json
Please accept/merge this PR and then add new records to it.

These are test cases.  ___Fix any that are wrong___ and add any new ones that you can think of.

We put the LCC code into our translator then it should return the corresponding list of classifiers.

The more different test cases that we have in this list, the more we can find and fix flaws in our translator.  https://github.com/cclauss/library-of-congress-classifier-hack